agent/requirements.txt
agno
openai
python-dotenv
uvicorn[standard]
fastapi